질문 작성 개발 및 카테고리 도메인 리팩터링 (#10)
* feat Question Service Repository 개발

* config Test하기 위해서 h2 dependency 추가 및 test/application.yaml 생성

* test Question Service Test

* modify Category ID BIGINT형으로 변경

* feat Question Save DTO, Question Converter 개발
Question과 Category 연관관계 맵핑부분이 힘들었음

* modify Question DTO Reponse 형식으로 변경
QuestionCategory 연관관계 추가

* modify QuestionCategory Entity Query에 ID 추가

* feat Category Response Dto, Category Converter 추가

+++ src/main/resources/conddae/dev/sql/카테고리_생성.sql
@@ -3,7 +3,7 @@ create database ccondae;
 use ccondae;
 
 create table category (
-    id int primary key,
+    id BIGINT primary key,
     count int default 0,
     name varchar(255)
 );
